Fix overflow checking for adcres=1 in multi-segment layout header.
The ugly bit of code here is to check whether the theoretical sample
range, specified by adcres and adczero, fits within an integer.  This
amounts to checking whether the calculation of msminv and msmaxv, in
the following lines, would overflow.

The calculation was done correctly (for single-segment records) in
check_segment, but was done wrong here.

The incorrect calculation gave spurious errors in the (weird) case of
a multi-segment record with layout adcres set to 1 (adcres=1 means the
signal has only two possible values, 0 and -1, so in that case dmax is
zero and this was undefined behavior).

This also would have failed to catch some erroneous cases (again, only
for multi-segment records.)